refactor: clean up organisation of pending-invoice code
